+++ app/backend/lib/reporting/gcpe.ts
@@ -1,6 +1,9 @@
 import writeXlsxFile, { Row } from 'write-excel-file';
 import { DateTime } from 'luxon';
-import { getConditionalApprovalDate } from '../../../lib/helpers/ccbcSummaryGenerateFormData';
+import {
+  getConditionalApprovalDate,
+  getFundingData,
+} from '../../../lib/helpers/ccbcSummaryGenerateFormData';
 import { performQuery } from '../graphql';
 import { HEADER_ROW, generateHeaderInfoRow } from './header';
 import columnOptions from './column_options';
@@ -17,9 +20,9 @@ import {
   handleProjectType,
   handleCbcEconomicRegions,
   handleCcbcEconomicRegions,
+  getCCBCFederalFundingSource,
 } from './util';
 import toTitleCase from '../../../utils/formatString';
-import { getFundingData } from '../../../lib/helpers/ccbcSummaryGenerateFormData';
 import { getFnhaValue } from '../dashboard/util';
 
 const getCbcDataQuery = `
@@ -32,14 +35,12 @@ const getCbcDataQuery = `
           projectNumber
           jsonData
           cbcByCbcId {
-            communitiesSourceDataByCbcProjectCommunityCbcIdAndCommunitiesSourceDataId {
-              edges {
-                cbcProjectCommunitiesByCommunitiesSourceDataId {
-                  nodes {
-                    communitiesSourceDataByCommunitiesSourceDataId {
-                      economicRegion
-                    }
-                  }
+            cbcProjectCommunitiesByCbcId(
+              filter: { archivedAt: { isNull: true } }
+            ) {
+              nodes {
+                communitiesSourceDataByCommunitiesSourceDataId {
+                  economicRegion
                 }
               }
             }
@@ -155,6 +156,7 @@ const getCcbcQuery = `
             analystStatus
             intakeNumber
             organizationName
+            program
             rowId
           }
         }
@@ -253,9 +255,7 @@ const generateExcelData = async (
       // economic region
       {
         value: handleCbcEconomicRegions(
-          node?.cbcByCbcId
-            ?.communitiesSourceDataByCbcProjectCommunityCbcIdAndCommunitiesSourceDataId
-            ?.edges
+          node?.cbcByCbcId?.cbcProjectCommunitiesByCbcId?.nodes
         ),
       },
       // federal funding source
@@ -360,7 +360,7 @@ const generateExcelData = async (
     );
     const row: Row = [
       // program
-      { value: 'CCBC' },
+      { value: node?.program },
       // announced by province
       {
         value: node?.applicationAnnouncedsByApplicationId?.nodes[0]?.announced
@@ -396,7 +396,7 @@ const generateExcelData = async (
         ),
       },
       // federal funding source
-      { value: 'ISED-UBF Core' },
+      { value: getCCBCFederalFundingSource(node) },
       // status
       { value: convertStatus(node?.analystStatus) },
       // project milestone complete percent